requirements.txt
# Data processing and analysis
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.1

# Machine Learning
scikit-learn==1.4.1
tensorflow==2.16.1
keras==2.16.1
lightgbm==4.1.0
xgboost==2.0.3
ta-lib==0.4.28 # Technical analysis library

# Web dashboard
dash==2.16.1
dash-bootstrap-components==1.5.0
plotly==5.19.0
flask==3.0.2

# Cryptocurrency API
ccxt==4.2.39  # Cryptocurrency exchange trading library
python-binance==1.0.19  # Binance API

# Utilities
joblib==1.3.2
tqdm==4.66.2
python-dotenv==1.0.1